+++ src/event.rs
@@ -188,6 +188,15 @@ impl<'a, N: Notify + 'a> input::ActionContext for ActionContext<'a, N> {
             Err(_) => warn!("Unable to start new Alacritty process: {} {:?}", alacritty, args),
         }
     }
+
+    fn toggle_fullscreen(&mut self) {
+        self.window_changes.toggle_fullscreen();
+    }
+
+    #[cfg(target_os = "macos")]
+    fn toggle_simple_fullscreen(&mut self) {
+        self.window_changes.toggle_simple_fullscreen()
+    }
 }
 
 /// The ActionContext can't really have direct access to the Window
@@ -196,17 +205,39 @@ impl<'a, N: Notify + 'a> input::ActionContext for ActionContext<'a, N> {
 /// the actual changes.
 pub struct WindowChanges {
     pub hide: bool,
+    pub toggle_fullscreen: bool,
+    #[cfg(target_os = "macos")]
+    pub toggle_simple_fullscreen: bool,
 }
 
 impl WindowChanges {
     fn clear(&mut self) {
         self.hide = false;
+        self.toggle_fullscreen = false;
+        #[cfg(target_os = "macos")]
+        {
+            self.toggle_simple_fullscreen = false;
+        }
+    }
+
+    fn toggle_fullscreen(&mut self) {
+        self.toggle_fullscreen = !self.toggle_fullscreen;
+    }
+
+    #[cfg(target_os = "macos")]
+    fn toggle_simple_fullscreen(&mut self) {
+        self.toggle_simple_fullscreen = !self.toggle_simple_fullscreen;
     }
 }
 
 impl Default for WindowChanges {
     fn default() -> WindowChanges {
-        WindowChanges { hide: false }
+        WindowChanges {
+            hide: false,
+            toggle_fullscreen: false,
+            #[cfg(target_os = "macos")]
+            toggle_simple_fullscreen: false,
+        }
     }
 }
 
@@ -281,6 +312,9 @@ pub struct Processor<N> {
     window_changes: WindowChanges,
     save_to_clipboard: bool,
     alt_send_esc: bool,
+    is_fullscreen: bool,
+    #[cfg(target_os = "macos")]
+    is_simple_fullscreen: bool,
 }
 
 /// Notify that the terminal was resized
@@ -326,6 +360,9 @@ impl<N: Notify> Processor<N> {
             window_changes: Default::default(),
             save_to_clipboard: config.selection().save_to_clipboard,
             alt_send_esc: config.alt_send_esc(),
+            is_fullscreen: false,
+            #[cfg(target_os = "macos")]
+            is_simple_fullscreen: false,
         }
     }
 
@@ -546,14 +583,43 @@ impl<N: Notify> Processor<N> {
             window.hide();
         }
 
-        if self.window_changes.hide {
-            window.hide();
+        #[cfg(target_os = "macos")]
+        {
+            use glutin::os::macos::WindowExt;
+
+            if self.window_changes.toggle_fullscreen && !self.is_simple_fullscreen {
+                self.set_fullscreen(window, !self.is_fullscreen);
+                self.is_fullscreen = !self.is_fullscreen;
+            }
+
+            if self.window_changes.toggle_simple_fullscreen && !self.is_fullscreen {
+                window.window().set_simple_fullscreen(!self.is_simple_fullscreen);
+                self.is_simple_fullscreen = !self.is_simple_fullscreen;
+            }
+        }
+
+        #[cfg(not(target_os = "macos"))]
+        {
+            if self.window_changes.toggle_fullscreen {
+                self.set_fullscreen(window, !self.is_fullscreen);
+                self.is_fullscreen = !self.is_fullscreen;
+            }
         }
 
         self.window_changes.clear();
         self.wait_for_event = !terminal.dirty;
 
         terminal
+    }
+
+    fn set_fullscreen(&self, window: &mut Window, fullscreen: bool) {
+        let glutin_window = window.window();
+        if fullscreen {
+            let current_monitor = glutin_window.get_current_monitor();
+            glutin_window.set_fullscreen(Some(current_monitor));
+        } else {
+            glutin_window.set_fullscreen(None);
+        }
     }
 
     pub fn update_config(&mut self, config: &Config) {